// src/types/registreSCRAP.ts
    export interface registreSCRAP {
        id_scrap: number;
        zone_affe_prob: string;
        annulation: string;
        desc_prob: string;
        cout_unit: string;
        machine: string;
        Desc_acmo: string;
        mini: string;
        table_elec: string;
        qnt_scrap: string;
        valeur_scrap: string;
        cause_prob: string;
        classification_cause: string;
        ac_immed_prend: string;
        rebut_remplacer: string;
        num_pcs_recup: string;
        qnt_rebF: string;
        h_interne: string;
        h_externe: string;
        cout_final: string;
        res_pos: string;
        ac_corr_suppl: string;
        N_ac_corr_ex: string;
        note: string;
        poids_rebut: string;
        id_ligne: string;
        compilateur: string
        id_lot: string;
        date_production: Date;
        id_article: string;
        odl_rep: string;
        new_odl: string;
        level: string
        oper_1: string;
        created_at:Date;
        N_pecRec:number;
    }